Remove protocols module
