+++ src/components/Loading.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import { AnimatePresence, motion } from 'framer-motion';
 import { Spinner } from '@codaco/ui';
 import { selectors as statusSelectors } from '@modules/ui/status';
-import { actionLocks as protocolsLocks } from '@modules/protocols';
+import { actionLocks as protocolsLocks } from '@modules/userActions';
 
 const variants = {
   hidden: { opacity: 0, transition: { delay: 0.5, duration: 0.5 } },
