+++ src/components/ProtocolControlBar.js
@@ -5,9 +5,8 @@ import { UnsavedChanges } from '@components/Dialogs';
 import { Button, Spinner } from '@codaco/ui';
 import { getProtocol } from '@selectors/protocol';
 import { getHasUnsavedChanges } from '@selectors/session';
-import { actionLocks as protocolsLocks } from '@modules/protocols';
 import { actionCreators as dialogActions } from '@modules/dialogs';
-import { actionCreators as userActions } from '@modules/userActions';
+import { actionCreators as userActions, actionLocks as protocolsLocks } from '@modules/userActions';
 import { selectors as statusSelectors } from '@modules/ui/status';
 import logoutIcon from '@app/images/home/log-out.svg';
 import ControlBar from '@components/ControlBar';
