+++ src/components/ProtocolStack.js
@@ -1,18 +1,17 @@
 import React from 'react';
-import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { get } from 'lodash';
 import { Flipped } from 'react-flip-toolkit';
-import { actionCreators as protocolsActions } from '@modules/protocols';
+import { actionCreators as userActions } from '@modules/userActions';
 import protocolCover from '@app/images/NC-File.svg';
 
 const getFilename = (path = '') => get(path.match(/([^/\\]+)$/), 1, path);
 
-const ProtocolStack = ({ unbundleAndLoadProtocol, protocol: { filePath } }) => (
+const ProtocolStack = ({ openNetcanvas, protocol: { filePath } }) => (
   <div
     className="protocol-stack"
-    onClick={() => unbundleAndLoadProtocol(filePath)}
+    onClick={() => openNetcanvas(filePath)}
   >
     <div className="protocol-stack__preview">
       <Flipped flipId={encodeURIComponent(filePath)}>
@@ -30,12 +29,12 @@ const ProtocolStack = ({ unbundleAndLoadProtocol, protocol: { filePath } }) => (
 
 ProtocolStack.propTypes = {
   protocol: PropTypes.object.isRequired,
-  unbundleAndLoadProtocol: PropTypes.func.isRequired,
+  openNetcanvas: PropTypes.func.isRequired,
 };
 
-const mapDispatchToProps = dispatch => ({
-  unbundleAndLoadProtocol: bindActionCreators(protocolsActions.unbundleAndLoadProtocol, dispatch),
-});
+const mapDispatchToProps = {
+  openNetcanvas: userActions.openNetcanvas,
+};
 
 export { ProtocolStack };
