+++ src/ducks/modules/root.js
@@ -4,12 +4,10 @@ import { combineReducers } from 'redux';
 import { combineEpics } from 'redux-observable';
 import { reducer as formReducer } from 'redux-form';
 import createTimeline from '../middleware/timeline';
-// import undoable from 'redux-undo';
 import app from './app';
 import dialogs from './dialogs';
 import guidance from './guidance';
 import protocol from './protocol';
-import protocols from './protocols';
 import recentProtocols from './recentProtocols';
 import session, { epics as sessionEpics } from './session';
 import settings from './settings';
@@ -39,7 +37,6 @@ export const rootReducer = combineReducers({
   guidance,
   locale: () => 'en-US',
   protocol: createTimeline(protocol, timelineOptions),
-  protocols,
   recentProtocols,
   session,
   settings,
