+++ src/utils/initFileOpener.js
@@ -1,19 +1,14 @@
-import React from 'react';
 import { ipcRenderer } from 'electron';
 import { store } from '@app/ducks/store';
-import { actionCreators as dialogActions } from '@modules/dialogs';
-import { actionCreators as protocolsActions } from '@modules/protocols';
-import { getActiveProtocol, getHasUnsavedChanges } from '@selectors/session';
-import { UnsavedChanges } from '@components/Dialogs';
+import { getActiveProtocol } from '@selectors/session';
+import { actionCreators as userActions } from '@modules/userActions';
 
-// TODO: robust save update
 const initFileOpener = () => {
   ipcRenderer.on('OPEN_FILE', (event, protocolPath) => {
     // eslint-disable-next-line no-console
     console.log(`Open file "${protocolPath}"`);
     const state = store.getState();
     const filePath = getActiveProtocol(state);
-    const hasUnsavedChanges = getHasUnsavedChanges(state);
 
     // If the protocol is already open, no op
     if (filePath === protocolPath) {
@@ -22,30 +17,7 @@ const initFileOpener = () => {
       return;
     }
 
-    if (!hasUnsavedChanges) {
-      // eslint-disable-next-line no-console
-      console.log('No unsaved changes, open.');
-      store.dispatch(protocolsActions.unbundleAndLoadProtocol(protocolPath));
-      return;
-    }
-
-    // eslint-disable-next-line no-console
-    console.log('Has unsaved changes, confirm.');
-
-    store.dispatch(dialogActions.openDialog(UnsavedChanges({
-      message: (
-        <p>
-          Attempting to open <em>{protocolPath}</em>,
-          but current protocol has unsaved changes.
-        </p>
-      ),
-      confirmLabel: 'Save changes and continue',
-    })))
-      .then((confirm) => {
-        if (!confirm) { return Promise.resolve(); }
-        return store.dispatch(protocolsActions.saveAndBundleProtocol())
-          .then(() => store.dispatch(protocolsActions.unbundleAndLoadProtocol(protocolPath)));
-      });
+    store.dispatch(userActions.openNetcanvas(protocolPath));
   });
 
   ipcRenderer.send('READY');
